requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
colorzero==2.0
cycler==0.11.0
fonttools==4.35.0
gpiozero==1.6.2
idna==3.3
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.2
opencv-contrib-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
torch==1.12.1
torchvision==0.13.1
typing-extensions==4.3.0
urllib3==1.26.11
